requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
Django==4.0.5
django-cors-headers==3.13.0
django-mongoengine==0.5.4
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
mongoengine==0.24.1
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
